// codeapp/src/generated/models/Cat_agentdetailsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cat_agentdetailsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cat_agentdetailsstatecode = keyof typeof Cat_agentdetailsstatecode;
export const Cat_agentdetailsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cat_agentdetailsstatuscode = keyof typeof Cat_agentdetailsstatuscode;

export interface Cat_agentdetailsBase {
  cat_agentdetail_name?: string;
  cat_agentdetailid: string;
  cat_agentid?: string;
  cat_agentname?: string;
  cat_billedcredit?: number;
  cat_channel?: string;
  cat_environmentid?: string;
  cat_environmentname?: string;
  cat_feature?: string;
  cat_knowledgesources?: string;
  cat_llmmodel?: string;
  cat_loadbatch?: string;
  cat_lookbackdays?: number;
  cat_nonbilledcredit?: number;
  cat_product?: string;
  cat_reportdate: string;
  cat_rowkey: string;
  cat_sourcedate?: string;
  cat_sourcefilename?: string;
  cat_tool?: string;
  cat_users?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cat_agentdetailsstatecode;
  statuscode?: Cat_agentdetailsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cat_agentdetails extends Cat_agentdetailsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}