// codeapp/src/generated/models/Cat_tenantcapacitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cat_tenantcapacitiesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cat_tenantcapacitiesstatecode = keyof typeof Cat_tenantcapacitiesstatecode;
export const Cat_tenantcapacitiesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cat_tenantcapacitiesstatuscode = keyof typeof Cat_tenantcapacitiesstatuscode;

export interface Cat_tenantcapacitiesBase {
  cat_allocated?: number;
  cat_asofdate?: string;
  cat_available?: number;
  cat_capacitytype?: string;
  cat_capturedon?: string;
  cat_consumed?: number;
  cat_entitled?: number;
  cat_paygoconsumed?: number;
  cat_rowkey?: string;
  cat_status?: string;
  cat_tenantcapacity_name?: string;
  cat_tenantcapacityid: string;
  cat_unit?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cat_tenantcapacitiesstatecode;
  statuscode?: Cat_tenantcapacitiesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cat_tenantcapacities extends Cat_tenantcapacitiesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}